@@ -1,34 +1,31 @@
 {
     "dependencies": {
+        "axios": "^1.7.7",
+        "bowser": "^2.11.0",
         "dayjs": "^1.11.13",
         "split-type": "^0.3.4",
-        "just-camel-case": "^6.2.0",
-        "object-hash": "^3.0.0",
-        "vue": "2.7.13",
-        "vue-observe-visibility": "^1.0.0",
-        "vue-resource": "1.5.3",
-        "vue-slide-up-down": "2.0.1",
-        "vue-ssr-carousel": "1.9.0",
-        "vue2-touch-events": "3.2.3",
-        "vuex": "3.6.2"
+        "swiper": "^11.1.14",
+        "vue": "3.5.12",
+        "vue3-slide-up-down": "^2.1.0"
     },
     "devDependencies": {
-        "@babel/core": "^7.25.2",
-        "@babel/preset-env": "^7.25.4",
+        "@babel/core": "^7.26.0",
+        "@babel/preset-env": "^7.26.0",
         "@symfony/webpack-encore": "4.7.0",
         "autoprefixer": "10.4.20",
-        "core-js": "3.38.1",
+        "core-js": "3.39.0",
         "css-mqpacker": "7.0.0",
         "dotenv-webpack": "^8.1.0",
         "file-loader": "6.2.0",
         "node-sass-glob-importer": "5.3.3",
         "postcss-loader": "7.0.2",
         "regenerator-runtime": "0.14.1",
-        "sass": "1.78.0",
+        "sass": "1.80.5",
         "sass-loader": "13.2.0",
         "url-loader": "4.1.1",
-        "webpack": "^5.94.0",
+        "webpack": "^5.95.0",
         "webpack-cli": "^5.1.4",
+        "webpack-dev-server": "^5.1.0",
         "webpack-notifier": "1.15.0",
         "wml-scaffold": "^0.0.2"
     },
